+++ core/src/lib/services/api.mjs
@@ -34,7 +34,10 @@ export const service = {
      * Lifecycle hook to determine whether to recreate the service
      * We just reuse the default hook here, telling it we need TLS configured.
      */
-    recreate: () => defaultRecreateServiceHook('api'),
+    recreate: () => {
+      ensureLocalPrerequisites()
+      return defaultRecreateServiceHook('api')
+    },
     /**
      * Lifecycle hook to determine whether to restart the service
      * We just reuse the default hook here, checking whether the service